Code for labeling CRBs moved to internal/backup.go
